+++ ExamenFinalSoftware/src/main/java/models/FabricaLg.java
@@ -10,30 +10,30 @@ package models;
  */
 public class FabricaLg implements FabricaElectrodomestico{
     @Override
-    public Lavadora crearLavadora(String modelo, String numSerie, String estado, double precio, double capacidad, boolean panel, int numProgramas){
+    public Lavadora crearLavadora(String modelo, String numSerie, String estado, int precio, double capacidad, boolean panel, int numProgramas){
         return (Lavadora) new LgLavadora(modelo, numSerie, estado, precio, capacidad, panel, numProgramas);
     }
     @Override
-    public Licuadora crearLicuadora(String modelo, String numSerie, String estado, double precio, int velocidades, double potencia, String tipoVaso, boolean picaHielo){
+    public Licuadora crearLicuadora(String modelo, String numSerie, String estado, int precio, int velocidades, double potencia, String tipoVaso, boolean picaHielo){
         return (Licuadora) new LgLicuadora(modelo, numSerie, estado, precio, velocidades, potencia, tipoVaso, picaHielo);
     }
     @Override
-    public Microondas crearMicroondas(String modelo, String numSerie, String estado, double precio,double capacidad, boolean panel, double potencia, boolean funcionGrill, boolean reloj, boolean cronometro){
+    public Microondas crearMicroondas(String modelo, String numSerie, String estado, int precio,double capacidad, boolean panel, double potencia, boolean funcionGrill, boolean reloj, boolean cronometro){
         return (Microondas) new LgMicroondas(modelo, numSerie, estado, precio, capacidad, panel, potencia, funcionGrill, reloj, cronometro);
     }
 
     @Override
-    public Refrigerador crearRefrigerador(String modelo, String numSerie, String estado, double precio, double capacidad, boolean dispensadorAgua, boolean dispensadorHielo, boolean alarmaPuertaAbierta) {
+    public Refrigerador crearRefrigerador(String modelo, String numSerie, String estado, int precio, double capacidad, boolean dispensadorAgua, boolean dispensadorHielo, boolean alarmaPuertaAbierta) {
         return (Refrigerador) new LgRefrigerador(modelo, numSerie, estado, precio, capacidad, dispensadorAgua, dispensadorHielo, alarmaPuertaAbierta);
     }
 
     @Override
-    public Stereo crearStereo(String modelo, String numSerie, String estado, double precio, int numUsb, int numParlantes, boolean karaoke, boolean ecualizador) {
+    public Stereo crearStereo(String modelo, String numSerie, String estado, int precio, int numUsb, int numParlantes, boolean karaoke, boolean ecualizador) {
         return (Stereo) new LgStereo(modelo, numSerie, estado, precio, numUsb, numParlantes, karaoke, ecualizador);
     }
 
     @Override
-    public Televisor crearTelevisor(String modelo, String numSerie, String estado, double precio, boolean bluetooth, boolean wifi, boolean controlVoz) {
+    public Televisor crearTelevisor(String modelo, String numSerie, String estado, int precio, boolean bluetooth, boolean wifi, boolean controlVoz) {
         return (Televisor) new LgTelevisor(modelo, numSerie, estado, precio, bluetooth, wifi, controlVoz);
     }
 }
