+++ ExamenFinalSoftware/src/main/java/models/FabricaPhilips.java
@@ -10,31 +10,31 @@ package models;
  */
 public class FabricaPhilips implements FabricaElectrodomestico{
     @Override
-    public Lavadora crearLavadora(String modelo, String numSerie, String estado, double precio, double capacidad, boolean panel, int numProgramas) {
+    public Lavadora crearLavadora(String modelo, String numSerie, String estado, int precio, double capacidad, boolean panel, int numProgramas) {
          return (Lavadora) new PhilipsLavadora(modelo, numSerie, estado, precio, capacidad, panel, numProgramas);
     }
 
     @Override
-    public Licuadora crearLicuadora(String modelo, String numSerie, String estado, double precio, int velocidades, double potencia, String tipoVaso, boolean picaHielo) {
+    public Licuadora crearLicuadora(String modelo, String numSerie, String estado, int precio, int velocidades, double potencia, String tipoVaso, boolean picaHielo) {
         return (Licuadora) new PhilipsLicuadora(modelo, numSerie, estado, precio, velocidades, potencia, tipoVaso, picaHielo);
     }
     @Override
-    public Microondas crearMicroondas(String modelo, String numSerie, String estado, double precio,double capacidad, boolean panel, double potencia, boolean funcionGrill, boolean reloj, boolean cronometro){
+    public Microondas crearMicroondas(String modelo, String numSerie, String estado, int precio,double capacidad, boolean panel, double potencia, boolean funcionGrill, boolean reloj, boolean cronometro){
         return (Microondas) new PhilipsMicroondas(modelo, numSerie, estado, precio, capacidad, panel, potencia, funcionGrill, reloj, cronometro);
     }
 
     @Override
-    public Refrigerador crearRefrigerador(String modelo, String numSerie, String estado, double precio, double capacidad, boolean dispensadorAgua, boolean dispensadorHielo, boolean alarmaPuertaAbierta) {
+    public Refrigerador crearRefrigerador(String modelo, String numSerie, String estado, int precio, double capacidad, boolean dispensadorAgua, boolean dispensadorHielo, boolean alarmaPuertaAbierta) {
         return (Refrigerador) new PhilipsRefrigerador(modelo, numSerie, estado, precio, capacidad, dispensadorAgua, dispensadorHielo, alarmaPuertaAbierta);
     }
 
     @Override
-    public Stereo crearStereo(String modelo, String numSerie, String estado, double precio, int numUsb, int numParlantes, boolean karaoke, boolean ecualizador) {
+    public Stereo crearStereo(String modelo, String numSerie, String estado, int precio, int numUsb, int numParlantes, boolean karaoke, boolean ecualizador) {
         return (Stereo) new PhilipsStereo(modelo, numSerie, estado, precio, numUsb, numParlantes, karaoke, ecualizador);
     }
 
     @Override
-    public Televisor crearTelevisor(String modelo, String numSerie, String estado, double precio, boolean bluetooth, boolean wifi, boolean controlVoz) {
+    public Televisor crearTelevisor(String modelo, String numSerie, String estado, int precio, boolean bluetooth, boolean wifi, boolean controlVoz) {
         return (Televisor) new PhilipsTelevisor(modelo, numSerie, estado, precio, bluetooth, wifi, controlVoz);
     }
 }
