+++ ExamenFinalSoftware/src/main/java/models/LgStereo.java
@@ -11,7 +11,7 @@ package models;
 public class LgStereo extends Stereo{
     
 
-    public LgStereo(String modelo, String numSerie, String estado, double precio, int numUsb, int numParlantes, boolean karaoke, boolean ecualizador) {
+    public LgStereo(String modelo, String numSerie, String estado, int precio, int numUsb, int numParlantes, boolean karaoke, boolean ecualizador) {
         this.marca="LG";
         this.modelo=modelo;
         this.numSerie=numSerie;
